added dag support
