+++ env.cfg
@@ -4,3 +4,5 @@ POSTGRES_USER=nplstudent
 POSTGRES_PASSWORD=newprolab
 POSTGRES_DB=de15_lab08
 PGDATA=/var/lib/postgresql/data/pgdata
+AWS_SECRET_ID=Y***E
+AWS_SECRET_KEY=Y***M
